chore: move county and district table model to `api/districts`
